lib: rename js source to lower snake_case
This commit renames all JavaScript source files in lib to lower
snake_case.

Backport-PR-URL: https://github.com/nodejs/node/pull/19969
PR-URL: https://github.com/nodejs/node/pull/19556
