+++ lib/_stream_readable.js
@@ -29,7 +29,7 @@ const Stream = require('stream');
 const { Buffer } = require('buffer');
 const util = require('util');
 const debug = util.debuglog('stream');
-const BufferList = require('internal/streams/BufferList');
+const BufferList = require('internal/streams/buffer_list');
 const destroyImpl = require('internal/streams/destroy');
 const errors = require('internal/errors');
 var StringDecoder;
