+++ lib/internal/modules/cjs/loader.js
@@ -49,8 +49,9 @@ module.exports = Module;
 
 // these are below module.exports for the circular reference
 const asyncESM = require('internal/process/esm_loader');
-const ModuleJob = require('internal/modules/esm/ModuleJob');
-const createDynamicModule = require('internal/modules/esm/CreateDynamicModule');
+const ModuleJob = require('internal/modules/esm/module_job');
+const createDynamicModule = require(
+  'internal/modules/esm/create_dynamic_module');
 
 function stat(filename) {
   filename = path.toNamespacedPath(filename);
