+++ lib/internal/modules/esm/loader.js
@@ -1,11 +1,12 @@
 'use strict';
 
 const errors = require('internal/errors');
-const ModuleMap = require('internal/modules/esm/ModuleMap');
-const ModuleJob = require('internal/modules/esm/ModuleJob');
-const defaultResolve = require('internal/modules/esm/DefaultResolve');
-const createDynamicModule = require('internal/modules/esm/CreateDynamicModule');
-const translators = require('internal/modules/esm/Translators');
+const ModuleMap = require('internal/modules/esm/module_map');
+const ModuleJob = require('internal/modules/esm/module_job');
+const defaultResolve = require('internal/modules/esm/default_resolve');
+const createDynamicModule = require(
+  'internal/modules/esm/create_dynamic_module');
+const translators = require('internal/modules/esm/translators');
 
 const FunctionBind = Function.call.bind(Function.prototype.bind);
