+++ lib/internal/modules/esm/module_map.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const ModuleJob = require('internal/modules/esm/ModuleJob');
+const ModuleJob = require('internal/modules/esm/module_job');
 const { SafeMap } = require('internal/safe_globals');
 const debug = require('util').debuglog('esm');
 const errors = require('internal/errors');
