+++ lib/internal/vm/module.js
