+++ lib/vm.js
@@ -209,4 +209,4 @@ module.exports = {
 };
 
 if (process.binding('config').experimentalVMModules)
-  module.exports.Module = require('internal/vm/Module').Module;
+  module.exports.Module = require('internal/vm/module').Module;
