+++ node.gyp
@@ -107,13 +107,13 @@
       'lib/internal/linkedlist.js',
       'lib/internal/modules/cjs/helpers.js',
       'lib/internal/modules/cjs/loader.js',
-      'lib/internal/modules/esm/Loader.js',
-      'lib/internal/modules/esm/CreateDynamicModule.js',
-      'lib/internal/modules/esm/DefaultResolve.js',
-      'lib/internal/modules/esm/ModuleJob.js',
-      'lib/internal/modules/esm/ModuleMap.js',
-      'lib/internal/modules/esm/ModuleWrap.js',
-      'lib/internal/modules/esm/Translators.js',
+      'lib/internal/modules/esm/loader.js',
+      'lib/internal/modules/esm/create_dynamic_module.js',
+      'lib/internal/modules/esm/default_resolve.js',
+      'lib/internal/modules/esm/module_job.js',
+      'lib/internal/modules/esm/module_map.js',
+      'lib/internal/modules/esm/module_wrap.js',
+      'lib/internal/modules/esm/translators.js',
       'lib/internal/safe_globals.js',
       'lib/internal/net.js',
       'lib/internal/os.js',
@@ -142,9 +142,9 @@
       'lib/internal/http2/util.js',
       'lib/internal/v8_prof_polyfill.js',
       'lib/internal/v8_prof_processor.js',
-      'lib/internal/vm/Module.js',
+      'lib/internal/vm/module.js',
       'lib/internal/streams/lazy_transform.js',
-      'lib/internal/streams/BufferList.js',
+      'lib/internal/streams/buffer_list.js',
       'lib/internal/streams/legacy.js',
       'lib/internal/streams/destroy.js',
       'lib/internal/wrap_js_stream.js',
