+++ test/message/esm_display_syntax_error_module.out
@@ -3,5 +3,5 @@ file:///*/test/fixtures/es-module-loaders/syntax-error.mjs:2
 await async () => 0;
 ^^^^^
 SyntaxError: Unexpected reserved word
-    at translators.set (internal/modules/esm/Translators.js:*:*)
+    at translators.set (internal/modules/esm/translators.js:*:*)
     at <anonymous>
